cut it down
